# springboot-httpbin/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: dyrnq@outlook.com
    url: https://github.com/dyrnq
  description: springboot-httpbin
  title: springboot-httpbin
  version: 0.0.1
servers:
- url: https://httpbin.org/
...
paths:
...
  /cookies/delete:
    get:
      parameters:
      - allowEmptyValue: true
        explode: true
        in: query
        name: freeform
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content: {}
          description: Redirect to cookie list
      summary: Deletes cookie(s) as provided by the query string and redirects to
        cookie list.
      tags:
      - Cookies
      x-accepts:
      - application/json
      x-tags:
      - tag: Cookies
  /cookies/set:
    get:
      parameters:
      - allowEmptyValue: true
        explode: true
        in: query
        name: freeform
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content: {}
          description: Redirect to cookie list
      summary: Sets cookie(s) as provided by the query string and redirects to cookie
        list.
      tags:
      - Cookies
      x-accepts:
      - application/json
      x-tags:
      - tag: Cookies
  /cookies/set/{name}/{value}:
    get:
      parameters:
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: value
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content: {}
          description: Set cookies and redirects to cookie list.
      summary: Sets a cookie and redirects to cookie list.
      tags:
      - Cookies
      x-accepts:
      - application/json
      x-tags:
      - tag: Cookies
...
  /status/{codes}:
    delete:
      parameters:
      - explode: false
        in: path
        name: codes
        required: true
        schema:
          type: string
        style: simple
      responses:
        "100":
          content: {}
          description: Informational responses
        "200":
          content: {}
          description: Success
        "300":
          content: {}
          description: Redirection
        "400":
          content: {}
          description: Client Errors
        "500":
          content: {}
          description: Server Errors
      summary: Return status code or random status code if more than one are given
      tags:
      - Status codes
      x-accepts:
      - application/json
      x-tags:
      - tag: Status codes
    get:
      parameters:
      - explode: false
        in: path
        name: codes
        required: true
        schema:
          type: string
        style: simple
      responses:
        "100":
          content: {}
          description: Informational responses
        "200":
          content: {}
          description: Success
        "300":
          content: {}
          description: Redirection
        "400":
          content: {}
          description: Client Errors
        "500":
          content: {}
          description: Server Errors
      summary: Return status code or random status code if more than one are given
      tags:
      - Status codes
      x-accepts:
      - application/json
      x-tags:
      - tag: Status codes
    patch:
      parameters:
      - explode: false
        in: path
        name: codes
        required: true
        schema:
          type: string
        style: simple
      responses:
        "100":
          content: {}
          description: Informational responses
        "200":
          content: {}
          description: Success
        "300":
          content: {}
          description: Redirection
        "400":
          content: {}
          description: Client Errors
        "500":
          content: {}
          description: Server Errors
      summary: Return status code or random status code if more than one are given
      tags:
      - Status codes
      x-accepts:
      - application/json
      x-tags:
      - tag: Status codes
    post:
      parameters:
      - explode: false
        in: path
        name: codes
        required: true
        schema:
          type: string
        style: simple
      responses:
        "100":
          content: {}
          description: Informational responses
        "200":
          content: {}
          description: Success
        "300":
          content: {}
          description: Redirection
        "400":
          content: {}
          description: Client Errors
        "500":
          content: {}
          description: Server Errors
      summary: Return status code or random status code if more than one are given
      tags:
      - Status codes
      x-accepts:
      - application/json
      x-tags:
      - tag: Status codes
    put:
      parameters:
      - explode: false
        in: path
        name: codes
        required: true
        schema:
          type: string
        style: simple
      responses:
        "100":
          content: {}
          description: Informational responses
        "200":
          content: {}
          description: Success
        "300":
          content: {}
          description: Redirection
        "400":
          content: {}
          description: Client Errors
        "500":
          content: {}
          description: Server Errors
      summary: Return status code or random status code if more than one are given
      tags:
      - Status codes
      x-accepts:
      - application/json
      x-tags:
      - tag: Status codes
...